// packages/wizzi.webapp/dist/src/services/wizzi/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi-js\dist\lib\artifacts\ts\module\gen\main.js
    package: wizzi-js@0.7.9
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi.apps\packages\wizzi.webapp\.wizzi\src\services\wizzi\types.ts.ittf
    utc time: Sat, 09 Jul 2022 08:31:38 GMT
*/
export type WizziConfigType = { 
    repoBasePath: string;
    metaBasePath: string;
    studioBasePath: string;
    jobsBasePath: string;
    crawlBasePath: string;
    githubBasePath: string;
    cheatsheetsBasePath: string;
    siteMetaBasePath: string;
    jsonDataMetaBasePath: string;
    storeKind: string;
    plugins: string[];
    pluginsBasePath: string;
};